GH-16 Refactor: update model interface
Since model objects are now only responsible for calculating the item
difficulties but not running the whole estimation process, the method
`run_estimation` is replaced by a new method `estimate_item_params`.
This new method takes a list of person params as input and returns a
list of item params.