GH-36 Save additional item parameters to DB
Each model can have more than a single item parameter - not just the "difficulty".
With this change, all the parameters are saved to the database after a
calculation of item parameters was triggered. Also, instead of just reading the
"difficulty" parameter from the DB,we now read all parameters from the DB.