GH-37 Use penalty to calculate question score
The penalty takes into account the last time a question was attempted. After 30
days, the penalty should be back to 0.
With this change, questions are ranked based on a score that takes both the fisher
information and the penalty into account.
If the penalty exceeds a certain threshold, it will never be selected. Right now,
questions that were attempted within the last 60 seconds meet this threshold.