GH-42 Refactor to facilitate combination of score functions
A test strategy involves calculation of different values, e.g.:
- calculate a penalty for a test item based on the last time that item was played
- calculate the fisher information for a test item
- calculate a score that takes both time penalty and fisher information into account
- etc...

To make it easier to add new functions to manipulate the score of a test item,
the following structure is now used:

1. An `item_score_modifier` interface was added. If you want to manipulate some
properties of a test item, you can implement this interface. Classes in the
`teststrategy\item_score_modifier` namespace will be loaded and called while fetching
the next question id.

2. The `item_score_modifier` classes might need some additional information to calculate
the new score. For example, to calculate the fisher information, we need to take the person
ability into account. Therefore, each `item_score_modifier` has to implement the
`get_required_context_keys()` method, which returns an array of strings that specify
the information needed by that class. The `update_score(array $context)` method
will receive a `$context` as its argument with the required information.

3. The context loader classes in the `local_catquiz\teststrategy\context\loader`
namespace can be used to populate the context.

4. These changes make the strategy classes very simple. They just have to return
an array of strings with the required score modifiers. The `teststrategy_base`
class takes care that all required data (context keys etc.) are present and calls
the score modifiers of the selected strategy in the given order.