GH-42 Refactor: use middlewares to update the score
Implements the middleware pattern, see e.g.
https://doeken.org/blog/middleware-pattern-in-php

With this change, we still have a chain of `item_score_modifier` classes to run
actions and calculate scores when we fetch the next question.
However, with this change it is now possible to return early from the chain and
skip the remaining item score modifications.