GH-41 Update SQL to show user- and total attempts
When retrieving the test items, instead of showing either to total number of
question attempts or the attempts for a single user, we now return both.
Same is true for the last question attempt time.

So in addition to the fields `attempts` and `lastattempttime`, which shows those
data for ALL users, the new fields `userattempts` and `userlastattempttime` shows
them for the given userid. If no userid was provided, those values will be `null`
and `0`, respectively.