GH-47 Refactor model hierarchy
Models that contact the R web service have to send all of the item responses at
once (the whole list), but the existing code called the submodels only on a per
-item basis.

With this change, all items are passed at once to the submodel classes.
To prevent changes to the existing raschbirnbaum[abc] classes, I added a
`model_raschmodel` abstract class that ensures the submodels are called with the
same data as before.