GH-77 Refactor to support non-dichotomous models
Before this change, the `catcalc` class checked the item response value and -
depending on its value (correct vs incorrect) - deferred calculation of the
likelihood or counter-likelihood to the model class.
With this change, the model class is passed the value of the item response. This
allows to add new models that support non-dichotomous outcomes.