GH-76 Refactor: move raschmodel outside subplugins
The only thing used in the `raschmodel.php` class was its `likelihood` function
to calculate the graphs of the item difficulties displayed in the UI.
This method was moved to the `model_raschmodel` class, which is already outside
the subplugins folder. Also, the method was renamed to `likelihood_1pl` to make
it explicit that it uses only the item difficulty.