GH-48 Itemparam calculation uses trusted regions
With this change, the trusted region methods  `restrict_to_trusted_region()`,
`get_log_tr_jacobian()` and `get_log_tr_hessian()` are now actually used.

In the `mathcat::newton_raphson_multi_stable()` method, we check if a call to
`restrict_to_trusted_region()` would change the value, which indicates that the
parameter values are in a critical region.
If so, the `get_log_tr_hessian()` and `get_log_tr_jacobian()` methods are added
to the existing callables.