GH-110 Refactor to use feedback generators
Currently, the structure is not very clear: some feedback strings are
returned from the teststrategy itself, while others are returned by the
`attempt_feedback` class. It could make sense to use a similar structure
like the pre-select task middleware: every strategy can specify a number
of feedback generators it wants to use. This would make it easy to
re-use existing generators for different strategies and to be able to
have a custom feedback for each strategy.
This is a first basic attempt to see if it would work.