GH-114 Tests: check item param calculation
Since I do not know what value should be calculated, the tests currently
just test for the value that is returned right now - so they all pass.

During calculation of the 3PL model, output is printed because the
estimated parameters are restricted to a trusted region. Because of this
output, PHPUnit marks the test as risky.