GH-85 Keep track of inactive questions and their fisher information
I assume that we have to take into account *all* questions of a subscale when
calculating that subscale's standarderror - even questions that were already
played and will therefore not be shown to the user anymore.
Since those questions are removed from the `$context['questions']` list, we now
keep track of the original list of questions in the `$context['original_questions]`
list and also include each question's fisher information there.