GH-125 Do not throw exceptions in any environment
The previous commit changed the handling of a missing last response
value in a way, that exceptions would only be thrown in development
environments. However, even there an exception leads to an endless loop
where the quiz attempt can not be finished. Therefore, with the current
change no exception will be thrown at all.