GH-144 Refactor feedback generators
Each feedback generator now has to implement a `load_data()` method, which collects
and returns all data required for this feedback.

When a quiz attempt is finished, we collect all data for the feedback and save it
to the database, before rendering the feedback from that data.

Adds a new public method `attemptfeedback->get_feedback_for_attempt($attemptid)`
to return the feedback for a given attempt.