GH-142 Fix logic and make breakinfo page more usable
With this change, we pass the timestamp of the end of the forced break as query
parameter to the breakinfo page, so that the user sees when he can continue.
The query parameter is only used to print the end time. Even if changed manually
by a user, this does not have any effect on the actual duration of the break.
Also, the id of the quiz is passed so that we can link back to the quiz.