GH-155 Only recalculate params when needed
Adds a new function that gets the number of new responses since a given timepoint.
Recalculation is now only performed if there is a new response associated to a
CAT context and CAT scales.
The event that informs about recalculations now gets a correct number of updated
items: a recalculation for an item is counted only once, even if its recalculated
by more than one model.