GH-171 Add test for moderate CAT
In this test strategy, questions with less attempts will be preferred to balance
out the number of attempts.
Refactors the `numberofgeneralattempts` class, so that the DB access can be overwritten
in a `numberofgeneralattempts_testing` child class.
Here we use fake data, so that for each question the number of attempts equals its
question ID.