Temporarily remove quiz break
This comments out or removes parts of the code that force a user to take a break
if the time to answer a question exceeds a certain threshold.
The functionality should be a bit different: no minimum time for the break
and some graphical indication of the time that is left to give a response.