GH-180 Fix missing after-attempt ability
The after-attempt ability is simply the new person ability for the
selected scale.
Before this change, the `personabilities` field of the `$attemptdata`
was set in the `comparetotestaverage` feedback generator. However, this
means that it won't be available if the selected teststrategy does not
use that feedback generator.
With this change, the `personabilites` is by default always added to the
attemptdata and independent of the feedback generators that are used.