GH-184 Configure feedback generators per strategy
This is more of a minimal example of how feedback generators can be
configured on a per-strategy basis: by passing parameters via the
constructor argument.
This is implemented here for the `customscalefeedback` class. If no
argument is given, feedbacks are sorted by ascending person ability.
However, this can be changed by passing a custom sort function, as for
the `infergreateststrength` strategy.