GH-204 Show ability before and after response
With this change, we display the ability before and after the respective question
in the graphical summary.
TODO: this should better be displayed in a table and not in a chart.

Note: Making `load_data()` return new data and using them in the `get_teacherfeedback()`
or `get_studentfeedback()` functions is problematic: when opening an old attempt,
the newly added data will be missing and this can cause errors when generating the
feedback.