Fix, Refactor: no cache for each catscale
Instead of caching each subscale under its own key, use the
`dataapi::get_all_catscales()` function to get all catscales. That
function already uses caching and this way we do not have to remember to
delete the cache key for a CAT scale when it is deleted, since that is
already handled by existing code.