Testing: update logic when to use testing classes
Some pre-select tasks have a CLASSNAME_testing.php alternative that is
useful to use when running unit tests. E.g. those classes can overwrite
methods that would usually access the database.

In order to allow more granular selection and not enable/disable ALL testing
classes at once, I changed the logic in strategy.php to check if a testing
class should be used.

Rename the updatepersonabilitytesting.php class to fit the testing class
naming scheme.

Also, the updatepersonability_testing.php class is modified to not throw
errors if no fake data are provided in the $context object passed to it.