Add quiz progress class
We need to keep track of data used by quiz attempts that are not
finished yet.
This was done in the cache previously but we also need to store data to
the DB in case the user logs out or the cache is cleared.
This class is responsible for storing and retrieving those data.