Fix: personability feedback check for empty data
Use `progress->get_playedquestions()` to check if there are any data to
be displayed instead of relying on `get_personabilities()` to be empty.

The reason that this change is required now is that the ability of the
main scale is now set by default. Therefore, `get_personabilities()`
will never be empty.