Re-add missing feedback table
In the table of played scales in the result details tab, a scale was
only shown if the 'numberofitemsplayed' key was set. I've unset this key
in a previous commit as part of a cleanup process to reduce the amount
of data stored in the database. This means that attempt feedbacks that
were generated with that code did not display the table.
In this commit, I've updated the logic in the mustache template to still
display scales, even if numberofitemsplayed is empty - just the
information about the number of played questions is hidden if this value
is not set.