Fix data retrieval for skipped question
If the last question was not answered, we did not get the last response
from the DB because the SQL query filtered for rows with fraction NOT
NULL. This is fixed here by filtering by state instead. To ensure that
the calculations work the same as before, the fraction is changed to 0.0
if the question state is `gaveup`.