GH-537 Do not overwrite fisherinfo
Somehow [1], in the debuginfo feedback generator, the fisherinformation is
overwritten. When this happens and the cache is deleted, the progress
class will read the fisherinformation from the database where it gets the
overwritten value. That value might now be a string like "NA" which
causes errors.

1: the fisherinformation is overwritten because it is part of the
`lastquestion` object in the `$newdata` array. Objects are passed by
reference, and therefore changing the value in the feedback generator
also changes it in outer scopes.