GH-508 Check hierarchy when importing items
This fixes importing items, when the scale hierarchy differs.
Previously, we checked just if the parent scale exists. With this
change, we check if it exists AND if the parent-child relationship is
correct.

For example, given the following data in the database:
Parent1 -> Scale 1 -> Subscale 1
Parent2 -> Scale 2 -> Subscale 2

When importing a new item to the following scale:
Parent1 -> Scale 2 -> Subscale 1
we have to not only check that those scales exist (they do),
but also if they exist in the given hierarchy. I.e., is Scale 2 a child
of Parent 1.
If not, create a new scale hierarchy.
This is fixed here.