GH-415 Implement reading multidimensional item params
Working with polytomous item parameters means that those parameters can
have a non-fixed form (e.g. arrays of different lengths). For example,
in the grmgeneralized model, the difficulty can be represented as an
array ([0.0 => 0.5, 0.5 => 0.6, 1.0 => 0.78]). The changes in this
commit instruct the model to return the item parameters in the required
form. Calling model_item_param::from_record($databaserecord) will ensure
the parameters are in the correct shape.