GH-649 Do not display feedback for hidden scales
In the allsubs test strategy, a scale is marked as hidden, if it does not meet the minimum requirements (SE or min questions).
Hidden scales are displayed in the result table to show how many questions were played. But we do not display ability values.
Also, with this change, we do not display feedback for hidden scales.