GH-639 Fix saving of feedback images
Images are loaded into the draft area and displayed in the html forms.
They are also displayed in shortcodes and the quiz feedback.

Note: when a quiz is saved for the first time, we do not have access to
the context module at the time when we want to save the iamges.
This means, that at the moment, images are only saved to the database
when an existing quiz is updated, but not when a quiz is first created.
We show a warning box to inform users about that.