GH-445 Make question feedback work everywhere
To make the question feedback work also when the feedback is displayed via shortcode or in the catmanager area, the attemptid and
slot are now directly passed as data-attributes to the HTML table fields.
Also, the webservice checks if question feedback is enabled for the given attemptid. If not, it returns just a string that it is
deactivated.