GH-393 Ensure models are calculated in correct order
When calculating values for Rasch-Birnbaum, we use the values calculated
by the Rasch model as start values. Furthermore, mixed Rasch-Birnbaum uses the
values calculated by Rasch-Birnbaum as startvalues.
This means we have to ensure that the calculations are done in a certain
order.
The model_sorter class does just that.