GH-393 Add characterization tests for item parameter calculations
These tests pin down the values that are now calculated so we can detect
unintended changes in the future.

It is possible to make the test output the data and write it to a CSV
file by setting the `CATQUIZ_CREATE_TESTOUTPUT` environment variable.
This will also create some output to stdout with some information about
durations of calculations.