GH-720 Fix import error with multiple question versions
This makes the code more robust: If there exist multiple question
versions, we try to remove older question versions from the scale it
should be assigned to. However, it is possible that the scale it should
be assigned to does not exist yet. This is accounted for in the updated
code.