GH-637 Update ability calculation if all answers are (in)correct
This change helps in estimating the ability if all given answers are
either correct or incorrect.
We will calculate the ability as before but in addition, also calculate
it under the assumption that the last answer is flipped (1 - $fraction).
We will then use the value that is more far from the parent ability.