GH-723 Update characterization tests for calculated item params
The matrix of responses is not correctly pruned to filter out items with
all-correct or all-incorrect responses. The same is done to filter out
persons or attempt with all-correct or all-incorrect answers.
This means the expected item parameters in the characterization test had
to be updated, because we previously didn't prune iteratively.