GH-791 Remove unused context key check
Since we are calling the preselect tasks directly via run now and do not
call the `process()` method in the abstract preselect_task class, the
`get_required_context_keys()` method is not called anymore.

Generally, this was not the best approach because its easy to use a
context key that is not listed in the `get_required_context_keys()`
method, so the actual code and the things defined there might diverge.

A better approach would be to use a class instead of an array for the
$context variable (e.g. preselect_context) and make sure it contains all
required fields as properties.